Put dict methods on the dict.prototype
I noted that `list` methods get assigned to `Array.prototype`, `str` methods get assigned to `String.prototype`, etc.
`dict` methods do not get assigned to `Object.prototype` (I am guessing this is for a good reason)

The current code has the `dict` methods being to each new instance in the constructor. It is tidier to declare them on the `dict` prototype and then create each dict instance from that.
This is also more efficient because properties placed on the dict prototype need only be declared once, as opposed to being declared for each dict.